Full fill all dataset functions
Made a dedicated query for dataset to retrieve cases table info on
related case_ids. Graphic representation of case info was added.
Example data file preview table was built and it helps user to
preview content of the example data. Related case info download
was added, and it's limited to cases table only.